+++ services/ui-service/src/shared/providers/fetch.provider.tsx
@@ -2,30 +2,26 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { ReactNode, useState } from "react"
 import { FetchInterceptor } from "@mswjs/interceptors/fetch"
+import Cookies from "js-cookie"
 
 const interceptor = new FetchInterceptor()
 
 interceptor.apply()
 interceptor.on("request", ({ request }) => {
-  request.headers.set(
-    "Authorization",
-    `Bearer ${localStorage.getItem("accessToken")}`
-  )
-  request.headers.set(
-    "refresh_token",
-    `${localStorage.getItem("refreshToken")}`
-  )
+  request.headers.set("Authorization", `Bearer ${Cookies.get("accessToken")}`)
+  request.headers.set("refresh_token", `${Cookies.get("refreshToken")}`)
 })
 
 interceptor.on("response", ({ response }) => {
   if (response.headers.has("token") || response.headers.has("Token")) {
     const newAccessToken =
       response.headers.get("token") ?? response.headers.get("Token") ?? ""
-    localStorage.setItem("accessToken", newAccessToken)
+    Cookies.set("accessToken", newAccessToken, { expires: 180 })
   }
 
   if (response.status === 401) {
-    localStorage.clear()
+    Cookies.remove("accessToken")
+    Cookies.remove("refreshToken")
     window.location.replace("/")
   }
 })